requirements.txt
foolbox==3.3.1
jupyter
pandas
pytorch-lightning==1.5.6
scikit-image
sklearn
seaborn
torchdiffeq==0.2.2